requirements.txt
streamlit==1.28.0
supabase==1.0.4
openai==0.28.1
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
python-dotenv==1.0.0
plotly==5.16.0
streamlit-authenticator==0.2.3
bcrypt==4.0.1
Pillow==10.0.0
openpyxl==3.1.2